% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countycensus.R
\name{countycensus}
\alias{countycensus}
\title{A function to pull and rework census data for transit calculations}
\usage{
countycensus(County)
}
\arguments{
\item{County}{An accepted county or city in Northern Virginia.}
}
\description{
This function allows you to easily pull census data to county level
and creates new variables.
Note: You must already have a Census API key set up. Load using readRenviron("~/.Renviron")
}
\examples{
countycensus("Arlington")
countycensus("Fairfax city")
countycensus("Fairfax County")
Nova counties and cities:
"Arlington", ,"Fairfax County", "Fairfax city", "Loudoun", "Alexandria City","Falls Church City"
}
